[FEAT] 상태 변경 메소드 Service로 위임
