+++ backend/src/main/java/com/back/domain/reservation/service/ReservationService.java
@@ -24,16 +24,19 @@ import com.back.global.s3.S3Uploader;
 import com.back.standard.util.page.PagePayload;
 import com.back.standard.util.page.PageUt;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ReservationService {
@@ -541,5 +544,32 @@ public class ReservationService {
                 totalAmount,
                 profileImgUrl
         );
+    }
+
+    @Transactional
+    public void updateReservationStatuses() {
+        updateClaimingStatus();
+        updatePendingRefundStatus();
+    }
+
+    private void updateClaimingStatus() {
+        List<Reservation> reservations = reservationRepository.findByStatus(ReservationStatus.CLAIMING);
+
+        reservations.forEach(reservation -> {
+            reservation.changeStatus(ReservationStatus.CLAIM_COMPLETED);
+            // Dirty Checking 작동
+        });
+
+        log.info("CLAIMING → CLAIM_COMPLETED 상태 변경 완료 - 처리 건수: {}", reservations.size());
+    }
+
+    private void updatePendingRefundStatus() {
+        List<Reservation> reservations = reservationRepository.findByStatus(ReservationStatus.PENDING_REFUND);
+
+        reservations.forEach(reservation -> {
+            reservation.changeStatus(ReservationStatus.REFUND_COMPLETED);
+        });
+
+        log.info("PENDING_REFUND → REFUND_COMPLETED 상태 변경 완료 - 처리 건수: {}", reservations.size());
     }
 }
